.eh_frame: CFA := (exp | regOffset)
Right now we were skipping over expressions but we weren't setting the
current CFAs that require a expression to be evaluated as such.

This PR doesn't tackle expression evaluation. It focuses on ensuring
that CFAs are marked as either a combination of register and an offset,
which we already support, or as an unsupported expression. In the latter
case we will not walk the stack in BPF.

This also fixes the table printing as rows with expressions had the
previous value, which is not correct.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>