bpf: Use 32 bit program counters
Use 32 bit program counters, halving the storage we need for program
counters. This reduces the compact info unwind rows by ~28%, making a
better use of the CPU caches.

This is achieved by storing the program counter relative to the lowest
PC in a chunk. Each chunk has 250k elements, so assuming that the
program counters don't have very big deltas, which is the case in every
binary we've tested, the difference fits in 32 bits.

Potential optimizations
=======================

Some other optimizations that we could do later on, once we measure their
impact and realize that they would be benefitial could be:

- Storing the program counters in a separate data structure, to minimise
  CPU cache misses while binary searching;
- Storing more unwind rows, now that the size is more compact. However,
  this might result in more CPU cache misses, and the binary search over
the table might have to increase its iteration count;

Test Plan
=========

Ran for several minutes, the profiles look good and zero errors.

```
         python3-7507    [010] d.h2.   260.459525: bpf_trace_printk: [[ stats for cpu 10 ]]
         python3-7507    [010] d.h2.   260.459526: bpf_trace_printk:    success=341
         python3-7507    [010] d.h2.   260.459527: bpf_trace_printk:    unsup_expression=0
         python3-7507    [010] d.h2.   260.459527: bpf_trace_printk:    unsup_frame=0
         python3-7507    [010] d.h2.   260.459527: bpf_trace_printk:    truncated=0
         python3-7507    [010] d.h2.   260.459528: bpf_trace_printk:    catchall=0
         python3-7507    [010] d.h2.   260.459528: bpf_trace_printk:    never=0
         python3-7507    [010] d.h2.   260.459528: bpf_trace_printk:    unknown_jit=108
         python3-7507    [010] d.h2.   260.459529: bpf_trace_printk:    total_counter=450
         python3-7507    [010] d.h2.   260.459529: bpf_trace_printk:    (not_covered=31)
```

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>